migrate discover session multiple data view test to discover (#213991)
Closes https://github.com/elastic/kibana/issues/191880

Skipped test had the following problems
* installed/uninstall sample data through the UI
* Generated new dashboard
* Tested discover session embeddable specific code.

This PR resolves these problems by
* Using already installed test data and avoiding UI flakiness of
installing/uninstalling sample data
* Uses pre-built dashboard
* moves test to discover app since its testing the discover session
embeddable and is not dashboard specific.

---------

(cherry picked from commit b7ce4e4f24ced5477a9b0b724e2bb534569b0fd6)
