+++ test/functional/apps/dashboard/group2/index.ts
@@ -31,6 +31,5 @@ export default function ({ getService, loadTestFile }: FtrProviderContext) {
     loadTestFile(require.resolve('./dashboard_filter_bar'));
     loadTestFile(require.resolve('./dashboard_filtering'));
     loadTestFile(require.resolve('./panel_expand_toggle'));
-    loadTestFile(require.resolve('./multiple_data_views'));
   });
 }
