+++ test/functional/apps/discover/embeddable/index.ts
@@ -23,5 +23,6 @@ export default function ({ getService, loadTestFile }: FtrProviderContext) {
     });
 
     loadTestFile(require.resolve('./_saved_search_embeddable'));
+    loadTestFile(require.resolve('./multiple_data_views'));
   });
 }
